Added solution classes
Added a class representing a single solution to our running dinner
optimization problem, as well as a class a single group of teams that
will cook a meal or be guests in that meal. The representation of a
solution is kept in a way where the solution can be changed via simple
permutation of indices and will always still stay a valid solution,
allowing for easy mutation of a solution to arrive at new versions of it
for our genetic algorithm.
Also added two unit tests for our new class