Added initializers
Implemented two initializer classes for the optimizer.
An initializer will produce an initial solution as startpoint for the
optimization.
The RandomInitializer will simply produce a solution at random, whereas
the FinalLocationInitializer will sort the cooking teams by their
distance to a final location already, making it an ideal start point if
the FinalLocationDistanceSolutionRater is used.