Added first DinnerPlan tests
Added two unit test for the dinner plan class. Also added an advanced
version of the Test.xlsx file to simulate a test with more teams. To be
able to run this new test offline without connection to the googleAPI,
added a dummy class that returns random distance values instead of
actually connecting to google.